refactor: modularize trade machine
